prototype/apps/backend/onboarding/src/logic/handle_quiz_results/requirements.txt
functions-framework
firebase-admin
google-cloud-firestore
google-cloud-aiplatform
vertexai